move git cacher to separate file
